# Improve hash-encoder performance - avoid unecessary Series construction