@PaulWestenthanner 2022-12-26
Hi,
thanks for your effort and this very detailed pull request. The errors in the pipeline are my fault (connected to #384). I'll try to sort it out this week and have your PR merged as well. Looks like a great improvement already