@bkhant1 2022-12-28
Just done it - the workflow needs approval to run.